requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
greenlet
asyncpg>=0.30.0
psycopg[binary]>=3.1.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
pydantic-settings>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Security & Auth
argon2-cffi
uuid-utils
email-validator
python-multipart
PyJWT

# Diagnostics & Tracking
asgi-correlation-id>=4.3.0

# HTTP & APIs
httpx>=0.27.0
requests
yfinance
ddgs
tweepy[async]

# Agentic Frameworks
langchain-core
langgraph
langchain-openai
langchain

# Data Processing
pandas
beautifulsoup4
readability-lxml
textstat

# Cache & Memory
redis>=7.1.0
langgraph-checkpoint-redis>=0.3.0
aiobreaker>=1.2.0

# Lifecycle & Tasks
apscheduler>=3.10.0
